:recycle: refactor: move tests to its own folder and fix folder name
